import: adjust file layout
