+++ pkg/ansible/dirs.go
@@ -1,0 +1,32 @@
+// Copyright 2020 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package ansible
+
+import (
+	"github.com/pingcap-incubator/tiops/pkg/meta"
+)
+
+// parseDirs sets values of directories of component
+func parseDirs(ins meta.InstanceSpec) (meta.InstanceSpec, error) {
+	switch ins.Role() {
+	case meta.RoleTiDB:
+	case meta.RoleTiKV:
+	case meta.RolePD:
+	case meta.RolePump:
+	case meta.RoleDrainer:
+	case meta.RoleMonitor:
+	case meta.RoleGrafana:
+	}
+	return ins, nil
+}
